refactor(daemon): split lifecycle and PTY modules
