+++ src/main/daemon/daemon-runtime-paths.ts
@@ -1,0 +1,30 @@
+import { existsSync, mkdirSync } from 'node:fs'
+import { join } from 'node:path'
+import { app } from 'electron'
+import { getDaemonLogFilePath } from '../observability/logs-directory'
+
+export function getDaemonRuntimeDir(): string {
+  const dir = join(app.getPath('userData'), 'daemon')
+  mkdirSync(dir, { recursive: true })
+  return dir
+}
+
+export function getDaemonHistoryDir(): string {
+  const dir = join(app.getPath('userData'), 'terminal-history')
+  mkdirSync(dir, { recursive: true })
+  return dir
+}
+
+export function getDaemonEntryPath(): string {
+  const appPath = app.getAppPath()
+  const basePath = app.isPackaged ? appPath.replace('app.asar', 'app.asar.unpacked') : appPath
+  const directEntryPath = join(basePath, 'daemon-entry.js')
+  return existsSync(directEntryPath)
+    ? directEntryPath
+    : join(basePath, 'out', 'main', 'daemon-entry.js')
+}
+
+export function getDaemonLogArgs(): string[] {
+  const disabled = (process.env.ORCA_DIAGNOSTICS_DISABLED ?? '').trim().toLowerCase()
+  return disabled === '1' || disabled === 'true' ? [] : ['--log-file', getDaemonLogFilePath()]
+}
