+++ src/main/daemon/daemon-server-shutdown.ts
@@ -1,0 +1,208 @@
+import type { Socket } from 'node:net'
+import { unlinkOwnedDaemonPidFile, unlinkOwnedDaemonTokenFile } from './daemon-spawner'
+import { DaemonServerSessionState } from './daemon-server-session-state'
+
+export abstract class DaemonServerShutdown extends DaemonServerSessionState {
+  async shutdown(): Promise<void> {
+    if (this.shutdownPromise) {
+      return this.shutdownPromise
+    }
+    const serverClose = this.beginOrdinaryShutdownFence()
+    this.shutdownPromise = this.finishOrdinaryShutdown(serverClose)
+    return this.shutdownPromise
+  }
+
+  protected beginOrdinaryShutdownFence(): Promise<void> {
+    this.idleShutdownState = 'shutting-down'
+    this.cancelInitialAdoptionTimer()
+    this.ordinaryShutdownServerClose ??= this.beginServerClose()
+    return this.ordinaryShutdownServerClose
+  }
+
+  protected async finishOrdinaryShutdown(serverClose: Promise<void>): Promise<void> {
+    this.unlinkOwnedEndpointArtifacts()
+    await this.disposeDaemonResources()
+    await serverClose
+  }
+
+  protected async finishRpcShutdown(serverClose: Promise<void>): Promise<void> {
+    await this.finishOrdinaryShutdown(serverClose)
+    this.onRpcShutdown()
+  }
+
+  protected unlinkOwnedEndpointArtifacts(): void {
+    unlinkOwnedDaemonTokenFile(this.tokenPath, this.token)
+    if (this.pidPath && this.launchNonce) {
+      unlinkOwnedDaemonPidFile(this.pidPath, process.pid, this.launchNonce)
+    }
+    this.ownedSocketIdentity = null
+  }
+
+  protected async disposeDaemonResources(): Promise<void> {
+    this.stopEndpointOwnershipWatch()
+    this.stopStreamBacklogProbe()
+    this.transientFactRelay.dispose()
+    this.cancelAllPendingPtySpawnPreparations()
+    try {
+      await this.host.dispose()
+    } catch (error) {
+      this.log.log('shutdown-dispose-failed', {
+        error: error instanceof Error ? error.message : String(error)
+      })
+    }
+    // Preserve canceled RPC replies across the extracted admission promise boundary.
+    await Promise.resolve()
+    this.streamDataBatcher.clear()
+    this.historySeedTransfers.dispose()
+    this.pendingShutdownReplies.clear()
+    for (const [, client] of this.clients) {
+      client.controlSocket.destroy()
+      client.streamSocket?.destroy()
+    }
+    this.clients.clear()
+    for (const socket of this.transportSockets) {
+      socket.destroy()
+    }
+    this.transportSockets.clear()
+  }
+
+  protected beginServerClose(): Promise<void> {
+    const server = this.server
+    this.server = null
+    if (!server) {
+      return Promise.resolve()
+    }
+    return new Promise<void>((resolve) => server.close(() => resolve()))
+  }
+
+  protected isIdle(): boolean {
+    if (this.createOrAttachInFlight > 0 || this.host.listSessions().length > 0) {
+      return false
+    }
+    if (this.endpointOwnershipLost) {
+      return true
+    }
+    return this.transportSockets.size === 0 && this.clients.size === 0
+  }
+
+  protected override reevaluateIdleShutdown(): void {
+    if (this.idleShutdownState !== 'running') {
+      return
+    }
+    if (this.retirementRequested) {
+      this.cancelInitialAdoptionTimer()
+      if (this.isIdle()) {
+        this.beginIdleShutdown()
+      }
+      return
+    }
+    if (!this.isIdle() || this.initialAdoptionDeadlineMs === null) {
+      this.cancelInitialAdoptionTimer()
+      return
+    }
+    if (this.initialAdoptionTimer !== null) {
+      return
+    }
+    const remainingMs = Math.max(0, this.initialAdoptionDeadlineMs - this.lifecycleClock.now())
+    if (remainingMs === 0) {
+      this.initialAdoptionDeadlineMs = null
+      this.retirementRequested = true
+      this.beginIdleShutdown()
+      return
+    }
+    this.initialAdoptionTimer = this.lifecycleClock.setTimeout(() => {
+      this.initialAdoptionTimer = null
+      this.initialAdoptionDeadlineMs = null
+      this.retirementRequested = true
+      this.beginIdleShutdown()
+    }, remainingMs)
+  }
+
+  protected armInitialAdoptionTimeout(): void {
+    this.initialAdoptionDeadlineMs = this.lifecycleClock.now() + this.initialAdoptionTimeoutMs
+    this.reevaluateIdleShutdown()
+  }
+
+  protected cancelInitialAdoptionTimer(): void {
+    if (this.initialAdoptionTimer === null) {
+      return
+    }
+    this.lifecycleClock.clearTimeout(this.initialAdoptionTimer)
+    this.initialAdoptionTimer = null
+  }
+
+  protected beginIdleShutdown(): void {
+    this.initialAdoptionTimer = null
+    if (this.idleShutdownState !== 'running') {
+      return
+    }
+    this.idleShutdownState = 'idle-shutdown-pending'
+    if (!this.isIdle()) {
+      this.idleShutdownState = 'running'
+      this.reevaluateIdleShutdown()
+      return
+    }
+    this.idleShutdownState = 'shutting-down'
+    const serverClose = this.beginServerClose()
+    this.shutdownPromise = this.finishIdleShutdown(serverClose)
+  }
+
+  protected async finishIdleShutdown(serverClose: Promise<void>): Promise<void> {
+    this.unlinkOwnedEndpointArtifacts()
+    await this.disposeDaemonResources()
+    await serverClose
+    this.onIdleShutdown()
+  }
+
+  protected recordFullyAuthenticatedDisconnect(wasFullyAuthenticated: boolean): void {
+    if (
+      !wasFullyAuthenticated ||
+      [...this.clients.values()].some((client) => client.authenticatedPairEstablished) ||
+      this.idleShutdownState !== 'running'
+    ) {
+      return
+    }
+    this.retirementRequested = true
+  }
+
+  protected shutdownReplyKey(clientId: string, requestId: string): string {
+    return `${clientId}\u0000${requestId}`
+  }
+
+  protected deferShutdownUntilReply(
+    clientId: string,
+    requestId: string,
+    socket: Socket,
+    finish: () => Promise<void>
+  ): void {
+    const key = this.shutdownReplyKey(clientId, requestId)
+    let started = false
+    let timer: ReturnType<typeof setTimeout>
+    const start = (): void => {
+      if (started) {
+        return
+      }
+      started = true
+      clearTimeout(timer)
+      socket.off('close', start)
+      socket.off('error', start)
+      this.pendingShutdownReplies.delete(key)
+      if (!this.shutdownPromise) {
+        this.shutdownPromise = finish()
+      }
+    }
+    timer = setTimeout(start, DaemonServerShutdown.SHUTDOWN_REPLY_FLUSH_TIMEOUT_MS)
+    timer.unref()
+    socket.once('close', start)
+    socket.once('error', start)
+    this.pendingShutdownReplies.set(key, { start })
+  }
+
+  protected stopEndpointOwnershipWatch(): void {
+    if (this.endpointOwnershipTimer === null) {
+      return
+    }
+    clearInterval(this.endpointOwnershipTimer)
+    this.endpointOwnershipTimer = null
+  }
+}
